rewrote lesson_1 homework to one class
